requirements.txt
altair==4.1.0
bokeh==2.2.3
numpy==1.19.5
pandas==1.2.0
psycopg2-binary==2.8.6
pydeck==0.5.0
SQLAlchemy==1.3.22
streamlit==0.74.1